Update lab 4
